% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{today_baseline}
\alias{today_baseline}
\title{The Treatment Options for Type II Diabetes in Adolescents and Youth (TODAY) study.}
\format{
We have three data sets that contain baseline demographics and longitudinal follow up for 699 participants:
\describe{
\item{ptid}{Patient identifier}
\item{txgrp}{Treatment Group}
\item{sex}{Sex}
\item{raceth}{Race/ Ethnicity}
\item{agebase}{Age at baseline}
}
}
\usage{
today_baseline
}
\description{
These are synthetic data sets inspired by the TODAY studY and its observational follow up.
}
\keyword{datasets}
