% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{today_fu}
\alias{today_fu}
\title{The Treatment Options for Type II Diabetes in Adolescents and Youth (TODAY) study.}
\format{
This dataset contains longitudinal follow up for 699 participants for the variables:
\describe{
\item{ptid}{Patient identifier}
\item{month}{Month of follow up}
\item{hba1c}{HbA1c (\%)}
\item{sbp}{Systolic Blood Pressure}
\item{dbp}{Diastolic Blood Pressure}
}
}
\usage{
today_fu
}
\description{
These are synthetic data sets inspired by the TODAY studY and its observational follow up.
}
\keyword{datasets}
