% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{today_lipids}
\alias{today_lipids}
\title{The Treatment Options for Type II Diabetes in Adolescents and Youth (TODAY) study.}
\format{
This dataset contains longitudinal follow up for 699 participants for the lipids and insulin variables:
\describe{
\item{ptid}{Patient identifier}
\item{month}{Month of follow up}
\item{cpep}{C-Peptide}
\item{ins}{Insulin}
\item{glu}{Glucose}
\item{chol}{Cholesterol}
\item{hdl}{HDL}
\item{trig}{Triglycerides}
}
}
\usage{
today_lipids
}
\description{
These are synthetic data sets inspired by the TODAY studY and its observational follow up.
}
\keyword{datasets}
